+++ client/src/services/club.js
@@ -1,5 +1,22 @@
 import { apiUrl } from "../utils";
 
+async function postClub(club) {
+  const response = await fetch(`${apiUrl}/clubs`, {
+    method: "POST",
+    credentials: "include",
+    body: club,
+    contentType: "application/json"
+  });
+
+  if (!response.ok) {
+    const errors = await response.json();
+    console.log(errors);
+    throw new Error(errors);
+  }
+
+  return response.json();
+}
+
 async function getClubs() {
   const response = await fetch(`${apiUrl}/clubs`, {
     credentials: "include"
@@ -9,4 +26,4 @@ async function getClubs() {
   return response.json();
 }
 
-export { getClubs };
+export { getClubs, postClub };
